Fix CodeQL bad upcast issues in this codebase

+++ core/include/vc/core/util/HashFunctions.hpp
@@ -11,7 +11,8 @@ namespace volcart
 
 /**
  * Hash for 3D vector types. Must support []-operator. It's designed for integer
- * types, so don't expect much if using with floating-point types.
+ * types, and only makes sense for positive integer values, so don't expect
+ * much if using with floating-point types.
  *
  * Uses the hash function from:
  * https://dmauro.com/post/77011214305/a-hashing-function-for-x-y-z-coordinates
@@ -24,7 +25,8 @@ struct Vec3Hash {
         std::size_t hash = (max * max * max) + (2 * max * v[2]) + v[2];
         if (max == v[2]) {
             auto val = std::max({v[0], v[1]});
-            hash += val * val;
+            hash +=
+                static_cast<std::size_t>(val) * static_cast<std::size_t>(val);
         }
         if (v[1] >= v[0]) {
             hash += v[0] + v[1];
